Feat: Added hateoas in client and bags routes

+++ src/presentation/controllers/ClientController.js
@@ -1,6 +1,4 @@
 import AppError from "../../infrastructure/errors/AppError.js";
-import jwt from "jsonwebtoken";
-import bcrypt from "bcrypt";
 
 class ClientController {
   constructor(clientService) {
@@ -9,21 +7,24 @@ class ClientController {
 
   async createClient(req, res, next) {
     try {
-      const client = await this.clientService.createClient(req.body);
-      return res.status(201).json(client);
+      await this.clientService.createClient(req.body);
+      return res.created();
     } catch (error) {
       next(error);
     }
-  }  async getClient(req, res, next) {
+  }  
+  
+  async getClient(req, res, next) {
     try {
-      const includeAddress = req.query.includeAddress === 'true';
-      const client = await this.clientService.getClient(req.params.id, { includeAddress });
+      let includeAddress = req.query.includeAddress;
+      includeAddress === true || 'true' ? includeAddress = true : includeAddress = false;
+      const client = await this.clientService.getClient(req.params.id, includeAddress);
       
       if (!client) {
         return next(new AppError('Cliente não encontrado', 'CLIENT_NOT_FOUND', 404));
       }
       
-      return res.json(client);
+      return res.hateoasItem(client);
     } catch (error) {
       next(error);
     }
@@ -31,21 +32,41 @@ class ClientController {
 
   async getAllClients(req, res, next) {
     try {
-      const includeAddress = req.query.includeAddress === 'true';
-      const clients = await this.clientService.getAllClients({ includeAddress });
-      return res.json(clients);
+      const { page, limit } = req.query;
+      const filters = {
+        includeAddress: req.query.includeAddress === 'true' ? true : false,
+        name: req.query.name || '',
+        email: req.query.email || '',
+        cpf: req.query.cpf || '',
+        phone: req.query.phone || '',
+        status: req.query.status || '',
+      }
+
+      const clients = await this.clientService.getAllClients(
+        page ? parseInt(page) : 1,
+        limit ? parseInt(limit) : 10,
+        filters
+      );
+
+      return res.hateoasList(clients.data, clients.pages);
     } catch (error) {
       next(error);
     }
-  }  async updateClient(req, res, next) {
+  }  
+  
+  async updateClient(req, res, next) {
     try {
+      req.params.id = req.user.entityId;
+      if (req.params.id !== req.user.entityId) {
+        return next(new AppError('Você não tem permissão para atualizar este cliente', 'UNAUTHORIZED', 403));
+      }
       const client = await this.clientService.updateClient(req.params.id, req.body);
       
       if (!client) {
         return next(new AppError('Cliente não encontrado', 'CLIENT_NOT_FOUND', 404));
       }
       
-      return res.json(client);
+      return res.ok(client);
     } catch (error) {
       next(error);
     }
@@ -53,39 +74,38 @@ class ClientController {
 
   async deleteClient(req, res, next) {
     try {
+      req.params.id = req.user.entityId;
+      if (req.params.id !== req.user.entityId) {
+        return next(new AppError('Você não tem permissão para deletar este cliente', 'UNAUTHORIZED', 403));
+      }
       await this.clientService.deleteClient(req.params.id);
-      return res.status(204).send();
+      return res.no_content();
     } catch (error) {
       next(error);
     }
-  }  async updateStatus(req, res, next) {
+  }  
+  
+  async updateStatus(req, res, next) {
     try {
+      if (parseInt(req.params.id) !== req.user.entityId) {
+        return next(new AppError('Você não tem permissão para atualizar o status deste cliente', 'UNAUTHORIZED', 403));
+      }
+
       if (req.body.status === undefined) {
         return next(new AppError('Status é obrigatório', 'MISSING_STATUS', 400));
       }
       
-      const client = await this.clientService.updateStatus(req.params.id, req.body.status);
+      const client = await this.clientService.changeClientStatus(req.params.id, req.body.status);
       
       if (!client) {
         return next(new AppError('Cliente não encontrado', 'CLIENT_NOT_FOUND', 404));
       }
       
-      return res.json(client);
+      return res.ok(client);
     } catch (error) {
       next(error);
     }
   }
-
-  async getActiveClients(req, res, next) {
-    try {
-      const includeAddress = req.query.includeAddress === 'true';
-      const clients = await this.clientService.getActiveClients({ includeAddress });
-      return res.json(clients);
-    } catch (error) {
-      next(error);
-    }
-  }
-
 }
 
 export default ClientController;
